refactor: move TestHTTPSConnectivity out of vm_helper.go
Move TestHTTPSConnectivity to its own connectivity_helper.go file since
it is a general network reachability helper, not VM-specific.
